adapter/requirements.txt
paho-mqtt==1.6.1
influxdb==5.2.3